agents/mami/pyproject.toml
[project]
name = "sye-agent-mami"
version = "0.1.0"
description = "Self-Improving Yolo Engine - Multi-model Classifier Agent"
requires-python = ">=3.11"
dependencies = [
  "smolagents[toolkit]>=1.22.0",
  "redis>=5.0,<7.0",
  "redisvl>=0.10.0",
  "openai>=2.6.1",
  "google-genai>=1.46.0",
  "anthropic>=0.71.0",
  "python-dotenv>=1.1.1",
  "neo4j>=5.22.0",
  "numpy>=2.3.4",
  "sentence-transformers>=5.1.2",
  "torch>=2.9.0",
  "rich>=14.2.0",
  "pydantic>=2.12.3",
]

[tool.ruff]
indent-width = 2

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
line-ending = "auto"
skip-magic-trailing-comma = false
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "I"] # pycodestyle errors, Pyflakes, isort
ignore = []
